+++ upkie/envs/backends/__init__.py
@@ -9,14 +9,12 @@
 """Backend implementations for Upkie environments."""
 
 from .backend import Backend
-from .genesis_backend import GenesisBackend
 from .mock_backend import MockBackend
 from .pybullet_backend import PyBulletBackend
 from .spine_backend import SpineBackend
 
 __all__ = [
     "Backend",
-    "GenesisBackend",
     "MockBackend",
     "PyBulletBackend",
     "SpineBackend",